refator: Funnel 템플릿 컴포넌트 shared 레이어에서 관리
