+++ src/features/sign-up/ui/index.ts
@@ -1,4 +1,2 @@
 export * from "./Header";
-
 export * from "./funnel/SignupFunnel";
-export * from "./funnel/FunnelStepPlate";
